ci: Upgrade to macOS v14.3.1 and Xcode v15.4.0
In the same spirit as the `next` branch did in 21a05d752cc140a685a971f63158
and in the spirit of trying to eliminate a very suspicious `libunwind`
error and is currently blocking the v1.58.0 release.

If this proves to work, we will need to commit to this version upgrade prior
to LTS to ensure our ability to best support macOS during the 1.x LTS
life-cycle.

Ref: https://www.apollographql.com/docs/graphos/reference/router-release-lifecycle
Ref: https://circleci.com/docs/using-macos/#supported-xcode-versions
Ref: https://github.com/apollographql/router/pull/5462
Ref: https://github.com/apollographql/router/pull/6298